Start tracking function dependencies and uses
The dynamic functions must be tracked with
respect to dependencies and usage. This is
very important whenever there are changes to
base functions (or functions used by other
functions). In its current form the remote
functions dependency is tracked but you
cannot see usage of the same. That is given
a remote function (say httpc:request/4)
which dynamic functions uses them.