you can write functions in efene as well
Yay! User can write dynamic functions in Efene as well.

Notice the special header for the same "#!efene"
Additionally note that comments must start with #_ and
be enclosed in double quotes.

Sample user program is as below, which can be run in the
websocket interfaces as ".run test_efene(1, 2)", where
the function is saved as ".save test_efene".

 #!efene
 #_ "first comment line shown in help"
 #_ "Another comment which must be in double quotes as well"
 fn
     case A, A: false
     case _, _: true
 end