you can write functions in efene as well
Awesome! You can now write dynamic functions in Elixir.

Notice the special header for the same "#!elixir"
Additionally note that comments must start with #

Sample user program is as below, which can be run in the
websocket interfaces as ".run test_elixir(1, 2)", where
the function is saved as ".save test_elixir".

 #!elixir
 # first comment line shown in help"
 # Another comment
 fn (a, b) ->
     a + b
 end