hugedata is used to store massive data
There is a requirement to store a lot of data by dynamic functions.
This use case would require storing a lot of data on disk,
which might be used as a temporary or permanent store.
Having said that this key category shall not be automatically
exported or importable due to its potential size per key.