saving function or whatis is across cluster
IMPORTANT: In case write fails to other nodes then
           it is not reattempted automatically,
           but instead the user is informed so that
           he/she can try again.

Note that deleting functions or purging them do not
delete automatically on all the other nodes.
This is primarily for protection to avoid
mistakes rather than anything else. In case the
user wants to delete function or whatis across
the cluster then it must be done in the cluster
nodes directly.