Two in-memory kv store fast write and seq write
There are two in memory kv store available to
the functions now; namely (a) concurrent fast write
and (b) sequential write. The objective of these two
variants is with respect to memory requirement and
write concurrency required. Note that in both of the
cases read is concurrent.

In case of concurrent fast write store, the memory
requirement is more compared to sequential write
store. In case you have a heavy write load then use
the concurrent fast write store, while pick the
sequential write when write concurrency requirements
is not too much.

A sample run is as follows:

    (beamparticle@127.0.0.1)1> beamparticle_fast_write_store:create(<<"a">>,
    1).
    true
    (beamparticle@127.0.0.1)2> beamparticle_fast_write_store:create(<<"a">>,
    1).
    false
    (beamparticle@127.0.0.1)3> beamparticle_fast_write_store:read(<<"a">>).
    {ok,1}
    (beamparticle@127.0.0.1)4> q().
    ok
    (beamparticle@127.0.0.1)5>