Compile and load python source once forever
It is much faster to compile given python source and load it
once forever, but then every subsequent load will keep
updating that, while adding more and more functions and
classes. Additionally, all the classes and functions are
in same namespace, so they can potentially call each other.
It is important to note that the name given "a2" while
calling __dynamic__ is just a placeholder to save the
source code within the python node. It can serve an
additional purpose of allowing "a2" to be directly
callable from the Erlang node, since it knows about it.

There is a key difference while invoking dynamic functions
in python; namely the arguments are passed as tuple rather
than list (because list can be confused with string in python
node). Hence the choice of using {M, F, A},
where M = <<"__dynamic__">>, F :: binary() and A :: {arg1, arg2, ...}

    (beamparticle@127.0.0.1)17> gen_server:call({'pythonserver',
    'python-1-beamparticle@127.0.0.1'}, {<<"MyProcess">>, <<"load">>, {<<"a2">>, <<"def
    add(a,b): return a+b\ndef a2(a,b): return add(a,b)">>}}).
    ok
    (beamparticle@127.0.0.1)18> gen_server:call({'pythonserver',
    'python-1-beamparticle@127.0.0.1'}, {<<"__dynamic__">>, <<"a2">>, {1,2}}).
    3