python node dies when erlang node exits
There is some trickiness with respect to non-blocking
read on sys.stdin in python. Other than that there is
an open case where when the Erlang process hosting
the python node (port) is killed (hard-kill) then
its terminate/2 is not invoked, hence it cannot
free up the allocated Id. FIXME!!! This results in
Id leaking and not bein reusable automatically.

There are number of solutions, one of them being
let another actor monitor and automatically release
Id, or write an actor which looks at allocated Ids
and tries to find associated process ownership.