maintain process reference with id
This is one way to track ownership, although there is
no common actor to keep track whether the python server
process is still alive and free the Id.

There is another TODO identified as part of this commit
wherein when the python node dies unexpectedly, then
it needs to be tested whether Erlang node gets an exit
signal or a message. I believe it is a exit signal but
then it needs to be tested.