pynode loads source and returns func arity
The function arity (i.e. number of arguments) is
returned by the pynode when the source is loaded.

A sample run is as shown below:

    gen_server:call({'pythonserver', 'python-1-beamparticle@127.0.0.1'},
                    {<<"MyProcess">>, <<"load">>,
                      {<<"a2">>,
                       <<"def add(a,b): return a+b\ndef a2(a,b): return add(a,b)">>}}).
    {ok,2}

    gen_server:call({'pythonserver', 'python-1-beamparticle@127.0.0.1'},
                    {<<"__dynamic__">>, <<"a2">>, {1,2}}).
    3

Note that even when the python source compiles, but do not contain
the function with given name then it gives {error, not_found}.
See below for an example. Note that there is no function "a3",
in the source code, while the source is a valid python source
and compiles.

    gen_server:call({'pythonserver', 'python-1-beamparticle@127.0.0.1'},
                    {<<"MyProcess">>,
                     <<"load">>,
                     {<<"a3">>,
                      <<"def add(a,b): return a+b\ndef a2(a,b): return add(a,b)">>}}).
    {error,not_found}