Python is partially integrated in beamparticle
There is a configuration in sys.config to automatically
start pool of actors for python nodes. This is
convenient rather than anything else.

Note that python code cannot be given via websocket yet,
because that would require some work with respect to
clean integration in beamparticle. Additionally, note that
the way python code can be loaded at present is via
the load interface in pynode, which must be invoked
whenever the node is started.