run single sub-worker in a java node
This is a simplistic approach wherein one JVM shall
serve only one request at a time, while other
requests shall be queued in the mailbox. If concurrency
is required then multiple JVMs must be started, so
tune higher num_workers in sys.config for javanode.
There are memory requirements for this case, so be aware
of the same although the workload is now more streamlined
and somewhat predictable performance or work guarantees are
provided. This does block JVM if the dynamic java function
is blocked on i/o, but at present that is a smaller
penalty to pay considering i/o must be performed in Erlang
node, while computationaly expensive task (non-i/o)
are done in java node.

TODO: limit memory when JVM is spun-up.