Java 8 can comple lambda from string
Java 8 allows compilation of lambda from string. There is
a great java 8 library which does that. This commit builds
on top of that functionality and allows usual methods
for loading, invoking and evaluating (without saving)
anonymous java class. Note that the class must have only
one method with the name "main", which will be used for
executing it. When evaluating the expression, then the
"main" public method (non-static) must have 0 arity, or
no arguments. Notice that import statements at the top
are automatically used as well when using other classes.

A sample anonymous class is as follows:

    () -> new Object(){
        public String main(){
            return "hello";
        }
    }

It is also possible to import stuff at the top as follows:

    import com.ericsson.otp.erlang.OtpErlangAtom;

    () -> new Object(){
        public OtpErlangAtom main(){
            return new OtpErlangAtom("ok");
        }
    }