JVM options are set in Erlang for more control
It is better to set JVM options while running the Erlang
port, so as to gain more control and allow changing of
the same dynamically.

Add a relative folder of "${cwd}"/javalibs where *.jar
files shall be automatically added to JVM classpath. So,
any jar present in "${cwd}"/javalibs shall be automatically
loaded by JVM on restart.