Python options are set in Erlang for more control
It is better to set Python options while running the Erlang
port, so as to gain more control and allow changing of
the same dynamically.

Add a relative folder of "${cwd}"/pythonlibs where *.zip
files shall be automatically added to PYTHONPATH. So,
any zip present in "${cwd}"/pythonlibs shall be automatically
loaded by Python on restart.