python node works flawlessly (so far)
The following python code when saved as test_python_simple_http

    #!python

    from pymongo import MongoClient
    import pandas as pd
    from datetime import datetime
    from datetime import timedelta
    import requests
    import sys

    from Pyrlang import term

    def main(post_binary, context_binary):
        return term.Binary('hello'.encode('utf-8'))

    def handle_event(data, context):
        return data

The regular raw functional invocation works as follows,
wherein test_python_simple_http.main() is invoked with
Erlang types. The result must be an Erlang type for
this to work.

$ curl --insecure -X POST \
  -d '{"arguments": ["hello", "world"]}' \
  -u 'root:root' -H 'content-type: application/json' \
  https://localhost:8282/fun/test_python_simple_http

{"speak":"","text":"","html":"<<\"hello\">>","json":{}}

While the simpler http invocation works as follows,
wherein the data (json) is passed as-is to the
function test_python_simple_http.handle_event(). Notice
that the data type of test_python_simple_http.handle_event()
are python str and the result must be a str as well.

$ curl --insecure -X POST -d '{"hello" : "world"}' \
  -u 'root:root' -H 'content-type: application/json' \
  https://localhost:8282/post/test_python_simple_http

{"hello": "world"}