lazy load python functions rather than all at once
There is no need for loading all the python functions
all at once. This both saves function datastore scan
and also solves the tricky multi-node cluster execution
scenarios, where not all the cluster nodes have
distributed the updated code to all the connected
python worker nodes.

This commit enhances the stacktrace in case of
exceptions as well, which goes to stderr and
eventually to the Erlang node console.

Be lazy and be happy!