lazy load java functions rather than all at once
There is no need for loading all the java functions
all at once. This both saves function datastore scan
and also solves the tricky multi-node cluster execution
scenarios, where not all the cluster nodes have
distributed the updated code to all the connected
java worker nodes.

Note that the java execution was not connected properly in
case of simple http interface, so this supposedly (untested)
fixes that as well.

Be lazy and be happy!