Write to stage then release to prod, Fixes #12
The intent is to save functions to stage area first and
then subsequently release to production once they are
ready. Notice that for Erlang, Elixir and Efene there are
two separate cashes maintained. One for prod version and
another for stage, so there is no runtime penalty here.
As for Java and Python each time there is a switch between
prod and stage version then the function is complied, so
this is very costly in case of flip flop.

TODO: optimize flip-flop for Java and Python in case
prod and stage versions are invoked very frequently.

Save a function test_stage with config/sys.config as follows:

    {code, [
         %% publish code changes to staging, while publish to
         %% production as a separate step
         {save_to_production, false}
     ]},

Function test_stage/2 is as follows:

    fun(Body, Context) ->
        <<"{\"msg\": \"hello there\"}">>
    end.

Request: default environment is prod, so test_stage is not found.

    curl --insecure -X POST -H 'content-type: application/json' -d '{}'
-u 'root:root' https://localhost:8282/post/test_stage

    {error,{throw,{error,<<"Please teach me what must I do with
test_stage(<<\"{}\">>,<<\"{}\">>)">>}}}

Request: Run function with stage environment, so the new function
exists.

    curl --insecure -type: application/json' -d '{}' -u 'root:root'
https://localhost:8282/post/test_stage?env=2

    {"msg": "hello there"}

Request: Use altenate endpoint to test in prod for stage function,
hence it fails.

    curl -X POST -H 'content-type: application/json' -d '{}'
http://localhost:8484/post/test_stage

    {error,{throw,{error,<<"Please teach me what must I do with
test_stage(<<\"{}\">>,<<\"{\\\"qs\\\": \\\"\\\"}\">>)">>}}}

Request: Use alternate endpoint to test stage function in stage
environment,
so it works.

    curl -X POST -H 'content-type: application/json' -d '{}'
http://localhost:8484/post/test_stage?env=2

    {"msg": "hello there"}