Allow dynamic html for nlp top page at port 8585
The web interface shipped along with BeamParticle is index-nlp.html
but many a times it may not be the what you are looking for.
In such situations instead of manually patching BeamPartile this
commit allows a dynamic function nlpfn_top_page/2 to serve
the custom response dynamically. See the following example
which sends out a simple html content. Needless to say
this dynamic function can only be implemented in Erlang, Elixir or
Efene which works directly on BEAM since the arguments to this
function may not map as-is to other programming languages.
Having said that Cowboy 2.0 works on map() heavily so
it just might work but not tested.

    #!erlang
    %% @doc dynamically generate html page for nlp interface
    %%
    %% fun(cowboy_req:req(), list()) -> {ok, {Body :: binary()
    %%                                        Headers:: map()}} |
    %%                                  {error, term()}.
    fun(Req0, Opts) ->
        Body = iolist_to_binary([<<"<html><head><title>Hello there</title>">>,
                                 <<"</head><body>Awesome! Dynamic html just">>,
                                 <<" works!</body></html>">>]),
        ResponseHeaders = #{},
        {ok, {Body, ResponseHeaders}}
    end.