Ensure that google oauth works for dev interface
This complete is part of #6 issue, but do not completely
close it because the redirection is shared between
the dev page (index.html) and nlp top page (index-nlp.html).
Since, these two would run on separate ports, so redirection
can only happen to single endpoint. This implies that
that the beamparticle_google_oauth_handler must behave
differently based on the Host and ensure that different
google client_id and secret are used. In summary there is
still some work to create two different google
config for client id and secret for index and index-nlp
page landings so that the redirection is accurate.