closes #28 capture stdout,stderr in py and java
TODO: This do not work predictably in Java for
some unknown reason. I suspect it is something to
do with flushing stdout and stderr streams.

Note that all you need to do is to dump logs to
stdout or stderr in Python or Java dynamic function
and then you can view it when running the function
via websocket dev interface. This will be very
useful for Python and Java devs.