closes #35 allow highperf /api/ for GET and POST
This is related to issue #34 and implements the GET
and POST for highperf interface, wherein the following works.

If you define a simple function test_get as follows:

    #!erlang
    fun(Body, Context) ->
        Body
    end.

Then the above is accessible via GET as follows, where
the query parameters are transformed into body (map).

    curl -v "http://localhost:8484/api/test_get?a=hello%20there"

Alternatively, the this also works

    curl -v -d '{"a": "hello there"}' \
      "http://localhost:8484/api/test_get"