closes #41 provide concurrent call semantics
One of the biggest advantages of using Erlang is massive concurrency, so
without a convenience hook to exploit it is a big miss. It is also
important to note that this pattern would be so common so as to require
in-built logic. Additionally, implementing it is non-trivial and prone
to errors.
Lets, add this so that developers can just invoke it and be happy.

Sample code for exploiting concurrency is as follows, where
add_two_numbers/2 and multiply_two_numbers/2 are already defined dynamic
functions. Notice that the third pattern uses the existing function
erlang:system_time/1.

    #!erlang
    fun() ->
        Tasks = [{add_two_numbers, [1,2]},
                       ,{multiply_two_numbers, [2, 4]}
                       {erlang, system_time, [second]}]
        ResultList = run_concurrent(Tasks, TimeoutMsec)
    end