closes #43 run async code concurrently
The fire-and-forget strategy gets a boost with this patch,
where it is now as easy as doing the following:

    #!erlang
    fun() ->
        Tasks = [{add_two_numbers, [1,2]},
                 ,{multiply_two_numbers, [2, 4]}
                 ,{erlang, system_time, [second]}],
        TimeoutMsec = 5000,
        JobRef = async_run_concurrent(Tasks, TimeoutMsec)
    end

The tasks are run asynchronously concurrently after a second,
while the result of the functions are not collected. In fact
the logs generated by these functions are not collected as well,
so debugging them is hard. Ensure that you test sequential and
sync concurrent via run_concurrent/2 first before using this
functionality.