[draft] fake inotify, parameterize git source path
There are still some places where the git source is
probably hard-coded or assumed to be at /opt/beamparticle-data/...
This needs to change otherwise things would not work when
the value in sys.config is anything different. The issue
is primarily with the new IDE and must be tested exhausitively.